+++ rpi/src/platform_rpi.cpp
@@ -6,6 +6,8 @@
 
 #include "platform.h"
 
+static bool s_isContinuousRendering = false;
+
 void logMsg(const char* fmt, ...) {
     va_list args;
     va_start(args, fmt);
@@ -14,16 +16,15 @@ void logMsg(const char* fmt, ...) {
 }
 
 void requestRender() {
-    // TODO: implement non-continuous rendering on RPi
+    bRender = true;
 }
 
 void setContinuousRendering(bool _isContinuous) {
-    // TODO: implement non-continuous rendering on RPi
+    s_isContinuousRendering = _isContinuous;
 }
 
 bool isContinuousRendering() {
-    return false;
-    // TODO: implement non-continuous rendering on RPi
+    return s_isContinuousRendering;
 }
 
 std::string stringFromResource(const char* _path) {
